+++ sample/src/main/java/com/synerise/sdk/sample/ui/dev/content/WidgetGridViewActivity.java
@@ -138,14 +138,12 @@ public class WidgetGridViewActivity extends BaseActivity {
         itemLayoutDetails.itemTitleStyle = Typeface.create("sans-serif-condensed", Typeface.NORMAL);
         itemLayoutDetails.itemTitleSize = 12;
         itemLayoutDetails.itemTitleColor = ContextCompat.getColor(Synerise.getApplicationContext(), R.color.charcoal);
-        itemLayoutDetails.itemTitleGravity = Gravity.LEFT;
         itemLayoutDetails.setItemTitleMargins(10, 0, 10, 0);
 
         //TextView Product price
         itemLayoutDetails.itemPriceStyle = Typeface.create("sans-serif-light", Typeface.BOLD);
         itemLayoutDetails.itemPriceSize = 12;
         itemLayoutDetails.itemPriceColor = ContextCompat.getColor(Synerise.getApplicationContext(), R.color.charcoal);
-        itemLayoutDetails.itemPriceGravity = Gravity.LEFT;
         itemLayoutDetails.setItemPriceMargins(10, 0, 10, 0);
 
         //TextView Product Sale price
@@ -277,14 +275,12 @@ public class WidgetGridViewActivity extends BaseActivity {
         itemLayoutDetails.itemTitleStyle = Typeface.create("sans-serif-condensed", Typeface.NORMAL);
         itemLayoutDetails.itemTitleSize = 12;
         itemLayoutDetails.itemTitleColor = ContextCompat.getColor(Synerise.getApplicationContext(), R.color.charcoal);
-        itemLayoutDetails.itemTitleGravity = Gravity.LEFT;
         itemLayoutDetails.setItemTitleMargins(10, 0, 10, 0);
 
         //TextView Product price
         itemLayoutDetails.itemPriceStyle = Typeface.create("sans-serif-light", Typeface.BOLD);
         itemLayoutDetails.itemPriceSize = 12;
         itemLayoutDetails.itemPriceColor = ContextCompat.getColor(Synerise.getApplicationContext(), R.color.charcoal);
-        itemLayoutDetails.itemPriceGravity = Gravity.LEFT;
         itemLayoutDetails.setItemPriceMargins(10, 0, 10, 0);
 
         //TextView Product Sale price